flask_app.py
import os
import re
import json
import logging
import asyncio
import httpx
import ccxt
from datetime import datetime
from flask import Flask, request
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
    level=logging.INFO
)
logger = logging.getLogger(__name__)

# Flask app
app = Flask(__name__)

# Telegram bot setup
TELEGRAM_TOKEN = os.getenv("TELEGRAM_BOT_TOKEN")
GROQ_API_KEY = os.getenv("GROQ_API_KEY")
CMC_API_KEY = os.getenv("CMC_API_KEY")
BITGET_API_KEY = os.getenv("BITGET_API_KEY")
BITGET_SECRET = os.getenv("BITGET_SECRET")

# ==================== DATA FETCHING FUNCTIONS ====================
# (Copy all your data fetching functions here: fetch_cmc_data, fetch_bitget_data, 
# fetch_data_for_coins, extract_coin_symbols, call_groq, and all command handlers)
# ...

# ==================== FLASK WEBHOOK ====================
@app.route(f"/{TELEGRAM_TOKEN}", methods=["POST"])
def webhook():
    """Handle incoming Telegram updates via webhook"""
    try:
        # Get the update from Telegram
        update_data = request.get_json()
        
        # Process the update asynchronously
        asyncio.run(handle_webhook_update(update_data))
        
        return "OK", 200
    except Exception as e:
        logger.error(f"Webhook error: {e}")
        return "Error", 500

async def handle_webhook_update(update_data):
    """Process the update and generate response"""
    update = Update.de_json(update_data, None)
    
    # Handle different update types
    if update.message:
        # Process message using your existing handle_message logic
        # You'll need to adapt your handle_message function to work here
        pass

# For local testing
if __name__ == "__main__":
    app.run()
