requirements.txt
python-telegram-bot==21.4
Flask==3.0.0
httpx==0.27.0
ccxt==4.3.0
python-dotenv==1.0.0
requests==2.31.0
